lo doy por terminado

+++ PECL2AyC/casilla.h
@@ -13,8 +13,8 @@
 using namespace std;
 
 struct casilla {
-    int valor;
-    int adj;
+    short valor;
+    short adj;
 };
 
 struct vecino {
